+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This project adheres to [semver](https://semver.org/)
 
+## [0.1.0-alpha] - 01-12-23
+### Added
+- Web client and server
+
+### Fixed
+- Throwing away scratch space refrence per database load.
+
 ## [0.0.5-alpha] - 01-06-23
 ### Added
 - a README